InstrumentManager/src/requirements.txt
fastapi
uvicorn
tinkoff-investments
